removed hashbits.{cc,hh}
