+++ khmer/_khmer.cc
@@ -48,7 +48,6 @@ Contact: khmer-project@idyll.org
 #include "khmer.hh"
 #include "kmer_hash.hh"
 #include "hashtable.hh"
-#include "hashbits.hh"
 #include "assembler.hh"
 #include "read_aligner.hh"
 #include "labelhash.hh"
