--- lib/hashbits.cc
@@ -1,67 +1,0 @@
-/*
-This file is part of khmer, https://github.com/dib-lab/khmer/, and is
-Copyright (C) 2010-2015, Michigan State University.
-Copyright (C) 2015-2016, The Regents of the University of California.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are
-met:
-
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-
-    * Redistributions in binary form must reproduce the above
-      copyright notice, this list of conditions and the following
-      disclaimer in the documentation and/or other materials provided
-      with the distribution.
-
-    * Neither the name of the Michigan State University nor the names
-      of its contributors may be used to endorse or promote products
-      derived from this software without specific prior written
-      permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-LICENSE (END)
-
-Contact: khmer-project@idyll.org
-*/
-#include <errno.h>
-#include <sstream> // IWYU pragma: keep
-
-#include "hashbits.hh"
-#include "hashtable.hh"
-#include "khmer_exception.hh"
-#include "read_parsers.hh"
-
-using namespace std;
-using namespace khmer;
-using namespace khmer:: read_parsers;
-
-void Hashbits::update_from(const Hashbits &otherBASE)
-{
-    if (_ksize != otherBASE._ksize) {
-        throw khmer_exception("both nodegraphs must have same k size");
-    }
-    BitStorage * myself = dynamic_cast<BitStorage *>(this->store);
-    const BitStorage * other;
-    other = dynamic_cast<const BitStorage*>(otherBASE.store);
-
-    // if dynamic_cast worked, then the pointers will be not null.
-    if (myself && other) {
-        myself->update_from(*other);
-    } else {
-        throw khmer_exception("update_from failed with incompatible objects");
-    }
-}
-
-// vim: set sts=2 sw=2:
