+++ lib/hashtable.hh
@@ -451,6 +451,14 @@ public:
     explicit CountingHash(WordLength ksize, std::vector<uint64_t> sizes)
         : Hashgraph(ksize, new ByteStorage(sizes)) { } ;
 };
+
+class Hashbits : public Hashgraph {
+public:
+    explicit Hashbits(WordLength ksize, std::vector<uint64_t> sizes)
+        : Hashgraph(ksize, new BitStorage(sizes)) { } ;
+
+    void update_from(const Hashbits &other);
+};
 }
 
 #define ACQUIRE_ALL_TAGS_SPIN_LOCK \
